Moved IMapChunk to Robust.Shared.Map. Moved Chunk class out of MapManager class.
