+++ Robust.Shared/Map/IMapChunkInternal.cs
@@ -1,0 +1,9 @@
+﻿using Robust.Shared.Timing;
+
+namespace Robust.Shared.Map
+{
+    internal interface IMapChunkInternal : IMapChunk
+    {
+        GameTick LastModifiedTick { get; }
+    }
+}
